[fix]: Changa writep data type to int

+++ 0x15-file_io/2-append_text_to_file.c
@@ -11,7 +11,7 @@ int append_text_to_file(const char *filename, char *text_content)
 {
 	int fd;
 	int len;
-	ssize_t writep;
+	int writep;
 
 	if (filename == NULL)
 		return (-1);
